Entities + tele colour only change in fixed function + below logic
Enities now store a colour, and that colour is set in the fixed render
function.

Graphics now stores 3 teleporter colours (inactive, active, and in
action) set in the fixed render function.

The variable render functions then render the entitity/teleporter with
the stored colour instead of generating a new colour. This means that
colours related to entities and teleporters should no longer update more
than 30 frames per second.

Non gameplay gamestates use a function to render an entity with a certain colour.
This is not an issue for now, but if those gamestate's render functions
were at some point in the future rendered at over 30 fps, this would be
an issue.

The fixed render functions are now called AFTER the logic functions
instead of before.
I did this because the first frame after entering a room, the fixed
render function had not been called and the variable render function was
being called, this resulted in sprites being rendered with the default
colour I set, black.
This should not cause any issues related to gameplay, it might even
reduce latency by one frame (or increase it), or it may break
everything, so I'm going to retest the entire game again.