Fixed player's old position not updating on treadmills/moving platforms.
Old position needed to be set before input was processed, so I removed
the old position being set in the entity update function, when the
player has died and when the game is at a complete stop and manually
placed the old position update code in main before input is processed,
this might break stuff, so I'll need to retest the entire game.

Tested final level, seems to work.